backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic[email]>=2.0.0
dnspython>=2.6.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
